pyproject.toml
[tool.poetry]
name = "agpt"
version = "0.5.0"
authors = ["Significant Gravitas <support@agpt.co>"]
readme = "README.md"
description = "An open-source attempt to make GPT-4 autonomous"
homepage = "https://github.com/Significant-Gravitas/AutoGPT/tree/master/autogpt"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
packages = [{ include = "autogpt" }]


[tool.poetry.scripts]
autogpt = "autogpt.app.cli:cli"
serve = "autogpt.app.cli:serve"


[tool.poetry.dependencies]
python = "^3.10"
autogpt-forge = { path = "../forge", develop = true }
# autogpt-forge = {git = "https://github.com/Significant-Gravitas/AutoGPT.git", subdirectory = "forge"}
click = "*"
colorama = "^0.4.6"
distro = "^1.8.0"
fastapi = "^0.109.1"
gitpython = "^3.1.32"
hypercorn = "^0.14.4"
openai = "^1.7.2"
orjson = "^3.8.10"
pydantic = "^2.7.2"
python-dotenv = "^1.0.0"
requests = "*"
sentry-sdk = "^1.40.4"

# Benchmarking
#agbenchmark = { path = "../benchmark", optional = true }
# agbenchmark = {git = "https://github.com/Significant-Gravitas/AutoGPT.git", subdirectory = "benchmark", optional = true}
pypdf2 = "^3.0.1"
reportlab = "^4.3.1"
pdfrw = "^0.4"
pillow = "^11.1.0"
google-cloud-vision = "^3.10.0"
matplotlib = "^3.10.1"
pikepdf = "^9.5.2"
pdf2image = "^1.17.0"
google-auth-oauthlib = ">=1.2.1" # Allow newer compatible versions
google-api-python-client = ">=2.163.0" # Allow newer compatible versions
google-auth-httplib2 = ">=0.2.0"
weasyprint = "^64.1"
pdfkit = "^1.0.0"
markdown-pdf = "^1.4"
python-docx = "^1.1.2"
docx2pdf = "^0.1.8"
markdown = "^3.7"
pdfplumber = "0.11.5"
pytesseract = "^0.3.13"
pymupdf = "^1.25.3"
opencv-python-headless = "^4.11.0.86"
openpyxl = "^3.1.5"
scikit-learn = "^1.6.1"
joblib = "^1.4.2"
shapely = "^2.0.7"
pdfminer-six = "20231228"
pyautogui = "^0.9.54"
pynput = "^1.8.0"
fillpdf = "^0.7.3"
plyer = "^2.1.0"
cryptography = "^44.0.2"
playwright = "^1.51.0"
rich = "^13.9.4"
pandas = "^2.2.3"

pyttsx3 = "^2.98"
sounddevice = "^0.5.1"
soundfile = "^0.13.1"
speechrecognition = "^3.14.2"
pyaudio = "^0.2.14"
livekit-agents = {version = "^1.0", extras = ["openai", "silero", "deepgram"]}
onnxruntime = "^1.21.1"
google-genai = "^1.14.0"
sseclient-py = "^1.8.0"


[tool.poetry.extras]
#benchmark = ["agbenchmark"]

[tool.poetry.group.dev.dependencies]
black = "^23.12.1"
flake8 = "^7.0.0"
isort = "^5.13.1"
pre-commit = "*"
pyright = "^1.1.364"

# Type stubs
types-colorama = "*"

# Testing
pytest = "*"
pytest-asyncio = "*"
pytest-cov = "*"
pytest-mock = "*"
pytest-recording = "*"
pytest-xdist = "*"

[tool.poetry.group.build]
optional = true

[tool.poetry.group.build.dependencies]
cx-freeze = { git = "https://github.com/ntindle/cx_Freeze.git", rev = "main" }
# HACK: switch to cx-freeze release package after #2442 and #2472 are merged: https://github.com/marcelotduarte/cx_Freeze/pulls?q=is:pr+%232442+OR+%232472+
# cx-freeze = { version = "^7.2.0", optional = true }


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.black]
line-length = 88
target-version = ['py310']
include = '\.pyi?$'


[tool.isort]
profile = "black"
skip_glob = ["data"]


[tool.pyright]
pythonVersion = "3.10"
exclude = ["data/**", "**/node_modules", "**/__pycache__", "**/.*"]
ignore = ["../classic/forge/**"]
